+++ petab_select/constants.py
@@ -41,6 +41,7 @@ MODEL_SUBSPACE_INDICES_HASH_MAP = (
     + string.ascii_uppercase
     + string.ascii_lowercase
 )
+PETAB_HASH_DIGEST_SIZE = None
 # If `predecessor_model_hash` is defined for a model, it is the ID of the model that the
 # current model was/is to be compared to. This is part of the result and is
 # only (optionally) set by the PEtab calibration tool. It is not defined by the
@@ -79,11 +80,14 @@ PARAMETERS = 'parameters'
 # PARAMETER_ESTIMATE = 'parameter_estimate'
 ESTIMATED_PARAMETERS = 'estimated_parameters'
 
+# Problem keys
 CRITERION = 'criterion'
 METHOD = 'method'
 VERSION = 'version'
 MODEL_SPACE_FILES = 'model_space_files'
+PROBLEM_ID = 'problem_id'
 
+CANDIDATE_SPACE = 'candidate_space'
 CANDIDATE_SPACE_ARGUMENTS = 'candidate_space_arguments'
 METHOD_SCHEME = 'method_scheme'
 PREVIOUS_METHODS = 'previous_methods'
@@ -92,6 +96,8 @@ PREDECESSOR_MODEL = 'predecessor_model'
 
 MODEL = 'model'
 MODELS = 'models'
+UNCALIBRATED_MODELS = 'uncalibrated_models'
+TERMINATE = 'terminate'
 
 # Parameters can be fixed to a value, or estimated if indicated with the string
 # `ESTIMATE`.